Completed all necessary styling on the page and fixed problems on accounts,community, forum and made event page dynamic

+++ Community/views.py
@@ -1,9 +1,12 @@
 from django.http import HttpResponse
-from django.shortcuts import render,redirect
+from django.shortcuts import render,redirect,get_object_or_404
 from django.db.models import Q
 from . models import Room,Topic,Message
 from . forms import RoomForm
 from django.contrib.auth.decorators import login_required
+from django.urls import reverse
+from django.contrib import messages
+
 # Create your views here.
 
 
@@ -26,14 +29,14 @@ def community(request):
 
     recent_activity = Message.objects.filter(Q(room__topic__name__icontains=q))
     
-    print(room_count)
+    # print(room_count)
     return render(request,"Community/community.html",{'rooms':rooms, 'topics':topics, 'room_count':room_count,'recent_activity':recent_activity})
 
 
 @login_required(login_url='')
 def room(request,pk):
     room = Room.objects.get(id=pk)
-    print(room)
+    print(room.host)
     room_messages = room.message_set.all().order_by('-created')
     participants = room.participant.all()
     
@@ -87,17 +90,47 @@ def updateRoom(request,pk):
     
     return render(request,'Community/room_form.html',context)
 
-@login_required(login_url='login')
-def deleteRoom(request,pk):
+
+# @login_required(login_url='login')
+# def deleteRoom(request, pk):
+#     room = get_object_or_404(Room, pk=pk)
     
-    room = Room.objects.get(id=pk)
+#     if request.user != room.host:
+#         return HttpResponse('You are not allowed here')
+
+#     if request.method == "POST":
+#         room.delete()
+#         return redirect('community')
+    
+#     return render(request, 'Community/delete.html', {'obj': room})
+
+
+
+@login_required(login_url='login')
+def deleteRoom(request, pk):
+    room = get_object_or_404(Room, pk=pk)
+    print("hello")
+    # print(request.user)
+    # print(room.host)
+    print("------------")
     if request.user != room.host:
         return HttpResponse('You are not allowed here')
+    
+    if request.method == 'POST':
+        print("hello")
+        
+        if request.user == room.host:
+            room.delete()
+            print("hello")
+            messages.success(request, 'Question deleted successfully.')
+            return redirect('community')
+        else:
+            messages.error(request, 'You are not authorized to delete this question.')
+            return redirect(reverse('room', kwargs={'pk': pk}))
+    
+    return redirect('home')  
 
-    if request.method == "POST":
-        room.delete()
-        return redirect('community')
-    return render(request,'Community/delete.html',{'obj':room})
+
 
 
 @login_required(login_url='login')
@@ -111,5 +144,4 @@ def deleteMessage(request,pk):
         message.delete()
         return redirect('community')
     return render(request,'Community/delete.html',{'obj':message})
-
